Revert "Enhance Candidate status updates"
